refactor: move all remotion related source code
