+++ app/_remotion/MyComp/NextLogo.tsx
