+++ app/_remotion/MyComp/Rings.tsx
