+++ app/_remotion/components/ProgressBar.tsx
