+++ app/_remotion/components/Tips.tsx
